prevent double clicking of start workflow button

+++ src/workflow/WorkflowManager.py
@@ -32,6 +32,9 @@ class WorkflowManager:
         Starts the workflow process and adds its process id to the pid directory.
         The workflow itself needs to be a process, otherwise streamlit will wait for everything to finish before updating the UI again.
         """
+        # Catch double presses of the button while app is in frozen state
+        if self.executor.pid_dir.exists(): return
+
         # Delete the log file if it already exists
         shutil.rmtree(Path(self.workflow_dir, "logs"), ignore_errors=True)
         # Start workflow process
